fix tarefa id counter - auto increment

+++ Spring/Exercicios/Exercicio-I/src/main/java/com/example/Exercicio_I/modules/controller/TarefasController.java
@@ -13,12 +13,14 @@ public class TarefasController {
 
     Database database = new Database();
 
+    //busca todas as tarefas
     @GetMapping("/tarefas/all")
     public List<Tarefa> getListaTarefas(){
         var list = List.copyOf(database.retrieveAll());
         return ResponseEntity.ok().body(list).getBody();
     }
 
+    //cria nova tarefa
     @PostMapping("/tarefas/create")
     public void createTarefa(@RequestBody CreateTarefaDTO dto){
         Tarefa tarefa = new Tarefa(dto.descricao());
@@ -27,7 +29,7 @@ public class TarefasController {
 
     //atualiza status (feito ou não feito)
     @PutMapping("/tarefas/{id}")
-    public void updateTarefa(@PathVariable("id") int id){
+    public void updateTarefa(@PathVariable("id") long id){
         for(Tarefa t: database.retrieveAll()){
             if(t.getId() == id){
                 if(t.isCompleta()){
@@ -39,6 +41,7 @@ public class TarefasController {
         }
     }
 
+    //deleta tarefa por id
     @DeleteMapping("/tarefas/{id}")
     public void deleteTarefa(@PathVariable int id){
         for(Tarefa t: database.retrieveAll()){
